[RENAME] : 폴더명 통일
